% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMM_GAR1_Deg_L.R
\name{ADMM.Deg.L}
\alias{ADMM.Deg.L}
\title{Estimate degree vector from a given Laplacian matrix}
\usage{
ADMM.Deg.L(
  L,
  rho = 0.01,
  epsilon = sqrt(1/nrow(S)),
  eps.abs = 1e-05,
  eps.rel = 0.001,
  max.iter = 1000,
  verbose = TRUE
)
}
\arguments{
\item{L}{A d by d normalized graph Laplacian}

\item{rho}{ADMM parameter (a positive number)}

\item{epsilon}{A positive small value}

\item{eps.abs}{ADMM convergence criterion}

\item{eps.rel}{ADMM convergence criterion}

\item{max.iter}{Maximum number of iterations to run the algorithm}

\item{verbose}{Trace of fitting procedure}
}
\value{
\itemize{
\item \code{v}: A d by 1 matrix
\item \code{w}: A d by 1 matrix
\item \code{deg}: A d by 1 matrix
\item \code{conv}: A boolean indicating whether the algorithm converged (TRUE) or didn't (FALSE).
}
}
\description{
\code{ADMM.Deg.L} returns an estimated degree vector to minimize ||Lv0||, with each element restricted to be at least epsilon.
}
