% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ADMM_Deg_L}
\alias{ADMM_Deg_L}
\title{ADMM for Degree estimation}
\usage{
ADMM_Deg_L(L, rho, epsilon, eps_abs, eps_rel, max_iter, verbose)
}
\arguments{
\item{L}{A d by d (normalized) graph laplacian}

\item{rho}{ADMM parameter (positive number)}

\item{epsilon}{small positive number}

\item{eps_abs}{ADMM stopping criterion}

\item{eps_rel}{ADMM stopping criterion}

\item{max_iter}{Maximum number of iterations}

\item{verbose}{Trace of algorithm}
}
\value{
A list object
\itemize{
\item \code{v}: A d by 1 matrix
\item \code{w}: A d by 1 matrix
\item \code{deg}: A d by 1 matrix
\item \code{conv} A boolean indicating convergence (TRUE if converged)
}
}
\description{
Given a (normalized) graph laplacian, estimate the degree vector by an ADMM algorithm.  Same as \code{ADMM.Deg.L}. but written in C++ and uses a different root-solving algorithm.
}
