% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAR_GAR_fit.R
\name{AutoCov}
\alias{AutoCov}
\title{Calculate sample auto-covariance matrix at lag=\code{order}}
\usage{
AutoCov(data, order = 0)
}
\arguments{
\item{data}{n by p data matrix}

\item{order}{nonnegative integer}
}
\value{
Returns a p-by-p symmetric matrix.
}
\description{
Calculates the sample autocovariance matrix at a specified lag/ordering.
}
