% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_fit.R
\name{GAR1.gf}
\alias{GAR1.gf}
\title{Goodness of Fit Test}
\usage{
GAR1.gf(
  S,
  nobs,
  lambda.v,
  rho.v = lambda.v,
  eps_thre = 1e-06,
  eps_abs = 1e-05,
  eps_rel = 0.001,
  max_iter = 10000,
  num.threads = 1
)
}
\arguments{
\item{`S`}{Estimate for covariance matrix, such as the MLE}

\item{`nobs`}{The number of observations used to calculate \code{S}}

\item{`lambda.v`}{Tuning parameter for GAR(1).  Positive number.}

\item{`rho.v`}{ADMM parameter.  Positive number.}

\item{`eps_thre`}{Small positive number.}

\item{`eps_abs`}{ADMM convergence criterion.}

\item{`eps_rel`}{ADMM convergence criterion.}

\item{`max_iter`}{Number of iterations to run initial fit on observed data}

\item{`num.threads`}{Number of threads to use for computing.}
}
\value{
p-value for the goodness of fit test
}
\description{
This function provides a goodness of fit test to see whether a GAR(1) model with the normalized laplacian is applicable.  It is valid for when the signal dimension is a most the number of observations available.
}
