% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMM_GAR1_fit_proc_package.R
\name{model_selec}
\alias{model_selec}
\title{Select tuning parameters for GAR(1) or TAR-GAR model}
\usage{
model_selec(resultList)
}
\arguments{
\item{resultList}{A list output from \code{GAR1_fit} or \code{TARGAR_fit}.}
}
\value{
A list object
\itemize{
\item \code{selected.model}: A list containing the ADMM output for the selected model.
\item \code{theta0}: The selected \code{theta0} estimate.
\item \code{theta1}: The selected \code{theta1} estimate.
\item \code{L}: The selected graph Laplacian estimate.
\item \code{v0}: The selected \code{v0} estimate (NULL when \code{step=2}).
\item \code{A.net.e}: A matrix encoding the (unweighted) graph chosen by the eBIC criterion.
\item \code{index}: Index for the optimal tuning parameters (lambda, net.thre) for the eBIC-selected model.
\item \code{lambda.v}: The selected \code{lambda.v} value.
\item \code{net.thre}: The selected \code{net.thre} value.
\item \code{ebic}: ebic score for the selected model.
\item \code{R_list}: For TAR-GAR fits, the selected AR filter matrices.
\item \code{eta}: For TAR-GAR fits, the selected polynomial filter coefficients.
\item \code{p}, \code{q}: For TAR-GAR fits, the AR order and polynomial order.
}
}
\description{
Given a fitted model path from \code{GAR1_fit} or \code{TARGAR_fit}, uses the eBIC
criterion to select the appropriate tuning parameters.
}
