\name{SGM-package}
\alias{SGM-package}
\alias{SGM}
\docType{package}
\title{
  Learning latent graphs from stationary signals
}
\description{
  This package offers a collection of functions to estimate a GAR(1) process, as proposed by Peng et al. in (2024).  The GAR(1) is the simplest of a collection of Spectral Graphical Models (SGM).  We provide a three-step estimation procedure for the model parameters, using an ADMM-based algorithm.  Additionally, we provide a model-selection procedure based on the eBIC criterion (Chen et al.) and a goodness of fit test. 
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Jedidiah Harwood, Jie Peng 

Maintainer: Jedidiah Harwood <jedharwood@ucdavis.edu>
}
\references{
  Peng et al. Learning Latent Graphs from Stationary Signals via a Graph Autoregressive Model (2024) 

  Isufi, E., Loukas, A., Perraudin, N., and Leus, G. (2019).
  Forecasting Time Series With VARMA Recursions on Graphs.
  \emph{IEEE Transactions on Signal Processing}, 67(18), 4870-4885.
  \doi{10.1109/TSP.2019.2929930}
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
