% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGM.R
\docType{_PACKAGE}
\name{SGM}
\alias{SGM}
\title{SGM}
\description{
A package for learning latent graphs from stationary signals.  This package provides 3 functions
}
\section{SGM functions}{

put them here
}

\references{
Isufi, E., Loukas, A., Perraudin, N., and Leus, G. (2019).
Forecasting Time Series With VARMA Recursions on Graphs.
\emph{IEEE Transactions on Signal Processing}, 67(18), 4870-4885.
\doi{10.1109/TSP.2019.2929930}
}
