% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAR_GAR_fit.R
\name{step.0}
\alias{step.0}
\title{Sample covariance of the innovation process}
\usage{
step.0(data, R1)
}
\arguments{
\item{data}{n by p matrix}

\item{R1}{p by p matrix}
}
\value{
A list containing:
\itemize{
\item \code{S}: A p by p sample covariance matrix of the innovation process
\item \code{theta0}: A positive number
}
}
\description{
Calculates the sample covariance of the innovation process, and the corresponding theta0 estimate, given a filter matrix R1
}
