% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAR_GAR_fit.R
\name{step.00}
\alias{step.00}
\title{Initial Estimate of Filter Matrix}
\usage{
step.00(data, eps.thre = 1e-06)
}
\arguments{
\item{data}{An n by p data matrix}

\item{eps.thre}{Threshold for pseudoinverse of Gamma0}
}
\value{
A p by p symmetrix matrix
}
\description{
Calculates an initial estimate of the graph filter matrix R1.
}
