% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAR_GAR_fit.R
\name{step.2}
\alias{step.2}
\title{Step 2 of TAR-GAR Fitting Procedure}
\usage{
step.2(data, L, theta0, eps = 1e-06)
}
\arguments{
\item{data}{An n by p data matrix}

\item{L}{A p by p (normalized) Laplacian Matrix}

\item{theta0}{A positive number}

\item{eps}{A (small) positive number}
}
\description{
Estimate the filter parameters (eta0, eta1) given (theta0, L), through a quadratic program with linear inequality constraints
}
