% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAR_GAR_fit.R
\name{step.3}
\alias{step.3}
\title{Step 3 of TAR-GAR Fitting Procedure}
\usage{
step.3(eta0, eta1, L)
}
\arguments{
\item{eta0}{A number}

\item{eta1}{A number}

\item{L}{A p-by-p matrix}
}
\description{
Update R1  given the estimated (eta0, eta1) and L parameters from steps 0-to-2.
}
