% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAR_GAR_fit.R
\name{step.lap.est}
\alias{step.lap.est}
\title{Refitting Step of TAR-GAR Procedure}
\usage{
step.lap.est(
  data,
  resList,
  A.net.e,
  lambda.v,
  net.thre,
  refit = 2,
  model = "LN",
  eps_thre = 1e-06,
  eps_abs = 1e-05,
  eps_rel = 0.001,
  max_iter = 50000,
  verbose = FALSE,
  time.vec
)
}
\arguments{
\item{data}{An n-by-p matrix}

\item{resList}{A list}

\item{A.net.e}{A p-by-p matrix}

\item{lambda.v}{A vector of positive numbers}

\item{net.thre}{A vector of (small) positive numbers}

\item{refit}{A positive integer}

\item{model}{A character}

\item{eps_thre}{A small positive number}

\item{eps_abs}{A small positive number}

\item{eps_rel}{A small positive number}

\item{max_iter}{A large integer}

\item{verbose}{A boolean}

\item{timevec}{A vector}
}
\description{
Given 0 pattern from net.thre, give an option to re-estimate L using GAR-strategy.  Takes in 0 pattern and results from the TAR-GAR passes, returns 3-step estimation procedure
}
