% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAR_GAR_fit.R
\name{step.thre}
\alias{step.thre}
\title{Thresholding Step}
\usage{
step.thre(resList, lambda.v, net.thre)
}
\arguments{
\item{resList}{Output from passes of TAR-GAR fitting procedure}

\item{lambda.v}{A vector of positive numbers}

\item{net.thre}{A vector of (small) positive numbers}
}
\description{
Use net.thre values to invoke sparsity in final pass estimate of L.  Takes in resList from final pass, and produces list of zero-patterns
}
