% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExampleTARGAR.R
\docType{data}
\name{targar}
\alias{targar}
\title{Example dataset: targar}
\format{
A list with the following components:
\describe{
\item{data}{A 100 observation by 100 node matrix with data simulated from an underlying TARGAR model, with parameters: \eqn{\theta_0 = 1}, \eqn{\theta_1 = 2}, \eqn{\eta_0 = 0.5}, \eqn{\eta_1 = 0.4 / \lambda_{\max}(L)}.}
\item{A.tr}{The true weighted adjacency matrix (100 by 100, with d = 100 nodes).}
\item{LN}{The true normalized Laplacian matrix (100 by 100, with d = 100 nodes).}
\item{theta0}{The true value of \eqn{\theta_0}, a positive number.}
\item{theta1}{The true value of \eqn{\theta_1}, a positive number.}
\item{eta0}{The true value of \eqn{\eta_0}.}
\item{eta1}{The true value of \eqn{\eta_1}.}
}
}
\source{
Simulated data from a TARGAR model
}
\usage{
data("targar")
}
\description{
A dataset generated from a TARGAR model.
}
\references{
Peng et al. (2024)
}
\keyword{datasets}
