% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMM_GAR1_Deg_L.R
\name{ADMM.Deg.L}
\alias{ADMM.Deg.L}
\title{Exported function}
\usage{
ADMM.Deg.L(
  L,
  rho = 0.01,
  epsilon = sqrt(1/nrow(S)),
  eps.abs = 1e-05,
  eps.rel = 0.001,
  max.iter = 1000,
  verbose = TRUE
)
}
\description{
Exported function
}
