% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ADMM_L2}
\alias{ADMM_L2}
\title{Added 10/9/2024}
\usage{
ADMM_L2(
  s,
  theta0,
  v,
  rho,
  lambda,
  model,
  Z_ini,
  W_ini,
  eps_thre,
  eps_abs,
  eps_rel,
  max_iter,
  verbose
)
}
\description{
Added 10/9/2024
}
