% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ADMM_L2}
\alias{ADMM_L2}
\title{ADMM algorithm for estimating (normalized) graph Laplacian given v0 and theta0}
\usage{
ADMM_L2(
  s,
  theta0,
  v,
  rho,
  lambda,
  model,
  Z_ini,
  W_ini,
  eps_thre,
  eps_abs,
  eps_rel,
  max_iter,
  verbose
)
}
\arguments{
\item{`s`}{an estimated covariance matrix, such as the sample covariance matrix}

\item{`theta0`}{A given graph filter parameter}

\item{`v`}{A given degree vector}

\item{`rho`}{ADMM parameter}

\item{`lambda`}{Tuning parameter}

\item{`model`}{A character specifying which type of Laplacian to use}
}
\description{
ADMM algorithm for estimating (normalized) graph Laplacian given v0 and theta0
}
