% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ADMM_L2_seq}
\alias{ADMM_L2_seq}
\title{Added 10/9/2024}
\usage{
ADMM_L2_seq(
  S,
  theta0,
  v0,
  Rho,
  Lambda,
  model,
  ini,
  eps_thre,
  eps_abs,
  eps_rel,
  max_iter,
  verbose
)
}
\description{
Added 10/9/2024
}
