% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ADMM_L2_Zero}
\alias{ADMM_L2_Zero}
\title{Added 10/09/2024}
\usage{
ADMM_L2_Zero(
  SS,
  theta0,
  v,
  rho,
  AA,
  model,
  Z_ini,
  W_ini,
  eps_thre,
  eps_abs,
  eps_rel,
  max_iter,
  verbose
)
}
\description{
Added 10/09/2024
}
