% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ADMM_Lap}
\alias{ADMM_Lap}
\title{Added 10/9/2024}
\usage{
ADMM_Lap(
  SS,
  V0,
  rho,
  lambda,
  model,
  ZZ_ini,
  WW_ini,
  phi_ini,
  eps_thre,
  eps_abs,
  eps_rel,
  max_iter,
  Z_max_iter,
  Z_conv_abs,
  Z_conv_rel,
  verbose
)
}
\description{
Added 10/9/2024
}
