% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ADMM_Lap_Zero}
\alias{ADMM_Lap_Zero}
\title{Added 10/09/2024}
\usage{
ADMM_Lap_Zero(
  SS,
  V0,
  rho,
  AA,
  model,
  ZZ_ini,
  WW_ini,
  phi_ini,
  eps_thre,
  eps_abs,
  eps_rel,
  max_iter,
  Z_max_iter,
  Z_conv_abs,
  Z_conv_rel,
  verbose
)
}
\description{
Added 10/09/2024
}
