% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMM_GAR1_fit_proc_package.R
\name{GAR1_fit}
\alias{GAR1_fit}
\title{Exported function}
\usage{
GAR1_fit(
  S,
  nobs,
  lambda.v,
  net.thre,
  model,
  step = 3,
  rho.v = lambda.v,
  eps_thre = 1e-06,
  eps_abs = 1e-05,
  eps_rel = 0.001,
  max_iter_1a = 10000,
  max_iter_2a = 10000,
  max_iter_3a = 10000,
  verbose = F
)
}
\description{
Exported function
}
