% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMM_GAR1_fit_proc_package.R
\name{GAR1_fit}
\alias{GAR1_fit}
\title{GAR(1) fitting procedure}
\usage{
GAR1_fit(
  S,
  nobs,
  lambda.v,
  net.thre,
  model,
  step = 3,
  rho.v = lambda.v,
  eps_thre = 1e-06,
  eps_abs = 1e-05,
  eps_rel = 0.001,
  max_iter_1a = 10000,
  max_iter_2a = 10000,
  max_iter_3a = 10000,
  verbose = F
)
}
\arguments{
\item{S}{An estimate of the covariance matrix, such as the MLE.}

\item{nobs}{The number of samples used to calculate \code{S}}

\item{lambda.v}{Tuning parameter to control sparsity of the estimated graph}

\item{net.thre}{Tuning parameter to control noisy entries in estimated graph}

\item{model}{\itemize{
\item \code{"LN"}: Fits a normalized graph Laplacian
\item \code{"L"}: Fits a graph Laplacian
\item \code{"LN.noselfloop"}: Fits a normalized graph laplacian assuming no self-loops.
}}

\item{step}{How many steps of the estimation procedure you want to run. Either 1, 2, or 3.}

\item{rho.v}{ADMM parameter (typically equal to \code{lambda.v})}

\item{eps_thre}{Small positive number}

\item{eps_abs}{ADMM convergence criterion}

\item{eps_rel}{ADMM convergence criterion}

\item{max_iter_1, }{max_iter_2, max_iter_3a Maximum number of iterations for algorithm}
}
\value{
A list object
\itemize{
\item \code{S} The p by p estimated covariance matrix
\item \code{result.L2.0} A list containing the Step 1a results
\item \code{result.0.post} A list containing the Step 2a results
\item \code{result.0S} A list containing the Step 3 results (NULL if \code{step<3})
\item \code{A.0.net} A matrix that encodes the estimated graph topology (NULL if \code{step<2})
\item \code{net.0.size} An integer containing the number of edges in the estimated graph (NULL if \code{step<2})
\item \code{v0.0.est} A p by 1 matrix containing the estimated degree vector (NULL if \code{step<3})
\item \code{theta0.0} A positive number. The estimated theta0 from Step 1 (and 2)
\item \code{theta0.0S} A positive number. The estimated theta0 from Step 3 (NULL if \code{step<3})
\item \code{conv.0.v0} A matrix containing convergence results for each combination of \code{lambda.v} (rows) and \code{net.thre} (columns) (NULL if \code{step<3})
}
}
\description{
\code{GAR1_fit} performs a three-step estimation procedure, using a penalized MLE approach, to estimate graph filter parameters \code{theta0} and \code{theta}, and the normalized graph Laplacian \code{L}.
}
\examples{
\dontrun{
### Load data
load("data/stocks.Rda")
n=nrow(stocks)
p=ncol(stocks)
model="LN"

### lambda and net.thre sequence
C.v=c(8,4,1)
lambda.v=C.v*sqrt(log(p)/n) 
rho.v=pmax(lambda.v, 0.01)

C.thre=exp(seq(log(1),log(0.1), length.out=12))
net.thre=C.thre*sqrt(log(p)/n) 

### Run GAR1_fit
S = var(stocks)*(n-1)/n
resList = GAR1_fit(S, n, lambda.v, net.thre, model, verbose = T,
                  max_iter_1a = 100000, max_iter_2a = 100000, max_iter_3a = 100000) 
}
}
