% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_fit.R
\name{GAR1_gf}
\alias{GAR1_gf}
\title{Goodness of Fit Test}
\usage{
GAR1_gf(
  S,
  nobs,
  lambda.v,
  rho.v = lambda.v,
  eps_thre = 1e-06,
  eps_abs = 1e-05,
  eps_rel = 0.001,
  max_iter = 10000,
  num.thread = 1,
  rep.boot = 100,
  seed = 1
)
}
\arguments{
\item{S}{Estimate for covariance matrix, such as the MLE}

\item{nobs}{The number of observations used to calculate \code{S}}

\item{lambda.v}{Tuning parameter for GAR(1). Positive number.}

\item{rho.v}{ADMM parameter. Positive number.}

\item{eps_thre}{Small positive number.}

\item{eps_abs}{ADMM convergence criterion.}

\item{eps_rel}{ADMM convergence criterion.}

\item{max_iter}{Number of iterations to run the Step 1 fit.}

\item{num.thread}{Number of threads to use for computing.}

\item{rep.boot}{Number of bootstrap samples to generate for the test.}

\item{seed}{Random seed used for reproducibility.}
}
\value{
p-value for the goodness of fit test
}
\description{
This function provides a goodness of fit test to see whether a GAR(1) model with the normalized Laplacian is applicable. It is valid when the signal dimension is at most the number of observations available.
}
