% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGM.R
\docType{package}
\name{SGM}
\alias{SGM-package}
\alias{SGM}
\title{SGM}
\description{
A package for learning latent graphs from stationary signals.  This package provides 3 functions
}
\section{SGM functions}{

put them here
}

