% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExample.R
\docType{data}
\name{stocks}
\alias{stocks}
\title{Stock data}
\format{
An object of class \code{matrix}
}
\usage{
data(stocks)
}
\description{
A 283 by 1007 matrix with Stock Data from the S&P 500.
}
\examples{
data(stocks)
}
\references{
Peng et al. 2024
}
\keyword{dataset}
