% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExample.R
\docType{data}
\name{stocks}
\alias{stocks}
\title{Example dataset: stocks}
\format{
A 283 by 1007 matrix
}
\source{
TBD
}
\usage{
data("stocks")
}
\description{
A matrix with Stock Data from the S&P 500.
}
\examples{
data(stocks)
}
\references{
Peng et al. 2024
}
\keyword{datasets}
