% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAR_GAR_fit.R
\name{TARGAR_fit}
\alias{TARGAR_fit}
\title{TAR-GAR Fitting Procedure}
\usage{
TARGAR_fit(
  data,
  lambda.v,
  net.thre,
  rho.v = lambda.v,
  num.pass = 2,
  model = "LN",
  eps.thre = 1e-06,
  eps_abs = 1e-05,
  eps_rel = 0.001,
  max_iter = 50000,
  refit = 2,
  verbose = F
)
}
\arguments{
\item{data}{An n-by-p matrix}

\item{lambda.v}{A vector of positive numbers.  These values are tuning parameters for the TAR-GAR procedure.}

\item{net.thre}{A vector of positive numbers.  These values are tuning parameters for the TAR-GAR procedure.}

\item{rho.v}{A vector of positive numbers.  The step size for the ADMM algorithm.  Set to \code{lambda.v} by default.}

\item{num.pass}{A positive integer.  Indicates how many passes to run before refitting. Three passes is recommended.}

\item{model}{A character.  Either \code{"LN"}, \code{"LN.noselfloop"}, or \code{"L"}.  Corresponds to GAR structure of the TAR-GAR innovation term.}

\item{eps.thre}{A small positive number.}

\item{eps_abs}{A small positive number.  Refers to ADMM stopping criteria. By default, set to 1e-5.}

\item{eps_rel}{A small positive number.  Refers to ADMM stopping criteria.  By default, set to 1e-3.}

\item{max_iter}{A large positive integer.  Refers to ADMM stopping criteria.  By default, set to 50000.}

\item{refit}{An integer.  Indicated number of steps in the GAR fitting procedure to use in the refitting step of the TAR-GAR procedure.  By default, set to 2 (indicates to use full GAR fitting procedure).}

\item{verbose}{A boolean.  Setting to TRUE prints out trace of the fitting procedure.}
}
\value{
A list containing:
\itemize{
\item \code{result.pass}: A list corresponding to the fitting results, pre-refitting.  Length is equivalent to \code{length(lambda.v)}.
\item \code{ini}: A list corresponding to the initial fitting results and first pass.  Length is equivalent to \code{length(lambda.v)}.
\item \code{refit}: A list corresponding to the final estimates using the TAR-GAR procedure.  Contains a collection of lists, with inner index corresponding to \verb{lambda index}, and outer index corresponding to \code{net.thre} index.
\item \code{A.net}: A list of matrices containing the zero-pattern of the estimated network.   Contains a collection of matrices, with inner index corresponding to \verb{lambda index}, and outer index corresponding to \code{net.thre} index.
\item \code{time.total}: Total time to run the TAR-GAR procedure on the specified sequence of tuning parameters.
}
}
\description{
Using the TAR-GAR fitting procedure, infers the (normalized) graph Laplacian matrix of a latent graph from temporally dependent data.  Fits to a sequence of tuning parameters (\code{lambda.v}, \code{net.thre}).
}
\examples{
\dontrun{
  ### Load data
  load("data/stocks.Rda")
  n=nrow(stocks)
  p=ncol(stocks)
  model="LN"
  
  ### lambda and net.thre sequence
  C.v=c(8,4,1)
  lambda.v=C.v*sqrt(log(p)/(n-1)) 
  rho.v=pmax(lambda.v, 0.01)
  
  C.thre=exp(seq(log(1),log(0.1), length.out=12))
  net.thre=C.thre*sqrt(log(p)/(n-1)) 
  
  ### Run GAR1_fit
  resList = TARGAR_fit(stocks, lambda.v, net.thre, rho.v, num.pass = 3)

}
}
